% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{pos}
\alias{pos}
\title{Return surface and pos results of MeCab by data.frame}
\usage{
pos(text, id_name = "id", dic = "", etc = "")
}
\arguments{
\item{text}{character vector.}

\item{id_name}{character. a name of a sentence id column.}

\item{dic}{character. a path to a user dictionary.}

\item{etc}{character. options to mecab.}
}
\value{
data.frame.
}
\description{
Return surface and pos results of MeCab by data.frame
}
\examples{
\dontrun{
pos(text,id_name="id",dic="",etc="")
}
}
