% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakati.R
\name{wakati}
\alias{wakati}
\title{Make "wakatigaki" to text}
\usage{
wakati(text, collapse = " ", dic = "", etc = "")
}
\arguments{
\item{text}{character vector.}

\item{collapse}{character. string to insert between each morphemes.}

\item{dic}{character. a path to a user dictionary.}

\item{etc}{character. options to mecab.}
}
\value{
character vector.
}
\description{
Make "wakatigaki" to text
}
\examples{
\dontrun{
wakati(text,collapse=" ",dic="",etc="")
}
}
