% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{pos}
\alias{pos}
\title{Return pos results of MeCab by data.frame}
\usage{
pos(text, type = c("surface", "base"), id_name = "id", dic = "", etc = "")
}
\arguments{
\item{text}{character vector.}

\item{type}{character. either "surface" or "base".}

\item{id_name}{character. a name of a sentence id column.}

\item{dic}{character. a path to a user dictionary.}

\item{etc}{character. options to mecab.}
}
\value{
data.frame.
}
\description{
pos() can return either surface form or base form of word and its pos information.
}
\examples{
\dontrun{
pos(text,type="surface",id_name="id",dic="",etc="")
}
}
